chore: remove dev/build artifacts, planning docs, and experiment files
Removed:
- Root build artifacts: RcppML_1.0.0.tar.gz, test_results.log, test_run.log, NEWS.md.bak
- Root planning docs: WORKQUEUE_*.md (10), CRAN_AUDIT.md, PRODUCTION_AUDIT.md, FACTOR_NET_VIGNETTE_PLAN.md
- experiments/ (35 files — research scripts and cached results)
- benchmarks/ (60+ files — benchmark harness, R scripts, results RDS)
- manuscript/ (40+ files — draft JSS paper LaTeX/figures/tables)
- logs/ (9 log files from dev sessions)
- tools/ one-off migration scripts (36 Python/R scripts; kept build/test scripts)
- vignettes/*.html and vignettes/*_files/ (pre-built outputs)
- vignettes/specs/ (15 planning docs — vignettes already written)
- tests/testthat/_problems/ (test failure artifacts)
- tests/cpp/ compiled objects (.o) and binary
- inst/extdata/ geo_sketch*.spz/gz, geo_sketch_meta.rds, pbmc3k_meta.rds (unreferenced)
- inst/build_config (auto-generated by configure)
- docs/dev/ (coverage matrix, hardening plan, workstreams)
- docs/ WORKQUEUE_*.{md,html}, copilot-instructions.{md,html}, PRODUCTION_AUDIT.{md,html}, CONTRIBUTING.{md,html}

+++ inst/include/FactorNet/nmf/nnls_streaming.hpp
@@ -51,7 +51,7 @@ namespace nmf {
  * @param threads      Number of threads (0 = auto)
  * @return H matrix (k x n)
  */
-template<typename Scalar = double>
+template<typename Scalar = float>
 Eigen::Matrix<Scalar, Eigen::Dynamic, Eigen::Dynamic> nnls_streaming_spz(
     const std::string& path,
     const Eigen::Matrix<Scalar, Eigen::Dynamic, Eigen::Dynamic>& W,
